client: initialize key repeat information
We can end up checking mRepeatKey.code before getting a key press is
recorded when the values are uninitialized.

This can happen if we get a key up for a previously held key or if key
repeats are disabled.

It's relatively harmless as it just stops a timer, but it can show up in
valgrind logs.

Change-Id: Ib7fe58b1976e9c89550c5a6252a34d66f83e2481
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>