Wayland plugins: replace header guards with #pragma once
Following qtbase/1c8884fc277c5916a420a3c14de68547a391f9fc's making
syncqt error out when an installable header contains #pragma once, we
can now flip the coin and use #pragma once as a static assertion that
the given header is not going to be installed, making it easier to
understand which headers are "public", which ones are "private" and
which ones are "neither, since never installed".

Do this for the Wayland platform support files, like
91f9ba2993a2652bcb349ce639aff67b5bfc8938 did for XCB.

Since these headers are neither public nor private, also remove "We
mean it" comments, where they existed.

Had to exclude the QtWlShellIntegration headers, because, for some
unknown reason, we install headers from the plugins/ subdir.

Pick-to: 6.10
Change-Id: I92f336f79d23134e96cbac13088dbd352e2d3a84
Reviewed-by: David Redondo <qt@david-redondo.de>