Client: implement toplevel icon protocol
This allows QWindow::setIcon to work on desktops that support the
standards. If an icon name is set, we assume it's from an installed
theme and provide the name, otherwise we provide buffers.

For SVGs the size hint from the compositor is used, otherwise we supply
pixmap data in all the sizes we have available.

[ChangeLog][Third-Party Code] New protocol synced from wayland-protocols

Change-Id: Iabbd928a3476aafa2d63a81e68299049e59991db
Reviewed-by: David Redondo <qt@david-redondo.de>