QByteArray: add conversion to std::string_view
96d67da420697cee10bdc537a1a592f6f22e2b8f didn't originally add this
conversion because on certain compilers (QNX) it's ambiguous. The
problem seems to be that QByteArray by default is convertible to const
char *; by adding a conversion to std::string_view, then the explicit
conversion

  std::string_view v(ba);

somehow becomes ambiguous.

Still, we've decided that we want the implicit conversion on non-broken
compilers, so this commits add it there.

[ChangeLog][QtCore][QByteArray] Added an implicit conversion operator
towards std::string_view.

[ChangeLog][QtCore][Important Behavior Changes] Converting a QByteArray
to a std::string_view will now create a view span over the entirety of
the byte array. Before, the view would have stopped at the first NUL in
the byte array (if there was one), unless the QT_NO_CAST_FROM_BYTEARRAY
macro was defined.

Change-Id: I70bbddac97636fbedb4d449eaa65cad4ecbdf0bf
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>