Client: Properly handle tablet manager being announced after wl_seat
When wl_seat is announced we create QWaylandInputDevice and there check
whether the tablet manager is available, create the tablet seat,
and store it in mTabletSeat.

If zwp_tablet_manager_v2 is announced after wl_seat, which
happens e.g. on Weston, we created the tablet seat in the
QWaylandTabletManagerV2 ctor, but didn't update mTabletSeat.

To address this, we move the handling to QWaylandDisplay and pass the
manager to QWaylandInputDevice, similar to how we do it e.g. for
text-input.

Pick-to: 6.8 6.7
Change-Id: I902375ec0085d34ba99cbe59bbbc4051e7d1e802
Reviewed-by: David Edmundson <davidedmundson@kde.org>
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>