Change handling of unspecified year in QCalendar::daysInMonth()
Previously, this was specified to return the usual number of days in
the month. What the date-time parser needs, though, is the longest
that the relevant month can be, in any year. So change its
specification to that. This affects all calendar backends, effectively
requiring daysInMonth() to treat Unspecified as if it were a leap
year.

[ChangeLog][QtCore][QCalendar][Important Behavior Changes] The
daysInMonth(month, Unspecified) corner case is now specified to return
the greatest number of days in that month across all years, where
previously it returned the usual number of days in the month. This
means QCalendar().daysInMonth(2) is now 29 rather than 28, with
similar for other calendars than the Gregorian default.

Change-Id: I165599c655df9fbd77700ed38d8860a4e5cca881
Reviewed-by: Marc Mutz <marc.mutz@qt.io>