Client: Implement frame-perfect resizing
With synchronous delivery we can be sure that any exposeEvent blocks
until the frame is rendered and committed.

This patch introduces a new override to make the render thread back off
starting a new frame when a resize is pending. If a configure event is
received we do not need the render thread to dispatch a new frame as the
main thread will do so explicitly when the configure event has been
processed.

To handle the case of a frame already being in flight at the time of the
main thread processing a configure event, this patch submits a new
exposeEvent with the old geometry. This will block until any existing
frames are fully submitted and the new exposeEvent is handled. Together
with the allowsIndependentThreadedRendering hook we can ensure that no
frame of the old size are going to be committed after we finish
processing and acking the incoming configure request.

Change-Id: Ic8ec9803800b17d2b8419da1a93e8dc388e702ab
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>