wasm: introduce QWasmSuspendResumeControl
QWasmSuspendResumeControl manages transfer of control from
JS to the wasm instance, when the wasm instance is asyncify
suspended.

Supported use cases include registerEventHandler()-type
event handlers, and also JS APIs which take a callback
functions, such as timers and requrestAnimationFrame.

The specific use cases are handled by adapters (not included
in this commit), which call the QWasmSuspendResumeControl API to
register JS/C++ event handler pairs:

  uint32_t index = control->registerEventHandler([](val){ ... });
  val jsHandler = control->jsEventHandler(index);

The C++ handler contains the user code, while the JS handler
can be passed to registerEventHandler or be used as a callback.

Change-Id: Ia1b1fd8884f0906759690dc7bc949c65a0248618
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>