MockCompositor: Protect against race condition
The QList m_fractionalScales was being changed from both the main thread
and the dispatch thread. This lead to read-after-free AddressSanitizer
errors when running autotests. Using a DirectConnection makes sure that
both appending and removal happens in the same thread. This does not
fix a theoretical race condition in DefaultCompositor::fractionalScale().

Fixes: QTBUG-130627
Pick-to: 6.8
Change-Id: Ie73f80d08929bcfcbd9940acdb4f86233df013cb
Reviewed-by: David Edmundson <davidedmundson@kde.org>