a11y macOS: Support titleUIElement property
In the macOS accessibility bridge, bridge the QAccessible::Label
relation to the corresponding titleUIElement NSAccessibility property [1]:

> An element that represents another element’s static text title (id)

In case there are multiple targets for the QAccessible::Label relation,
just report the first one.

With this in place, Accessibility Inspector on macOS now e.g. shows
"Locale (text) [QMacAccessibilityElement]" as the "Title UIElement"
of the corresponding combobox it labels in the calendarwidgets example
as expected and VoiceOver now says e.g. "English/Germany Locale General
Options, menu button" instead of just the current combobox value and
role "English/Germany, menu button", so the context and semantics of
the combobox becomes clearer to the user.

[1] https://developer.apple.com/documentation/appkit/nsaccessibility-swift.struct/attribute/titleuielement

Change-Id: I814658628c481fe71f17853f71dc5690971507b1
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>