Use ICU from Windows SDK for codec support (QStringConverter)
Windows has shipped a (limited) subset of ICU for quite a while:

  https://learn.microsoft.com/en-us/windows/win32/intl/international-components-for-unicode--icu-

The C API for this subset uses non-standard names for headers and lib
though, and is therefore not picked up by the ICU configure logic in
CMake.

This patch adds a separate configure test that enables the use of this
ICU API, specifically for text codec support in QStringConverter.
If a 'proper' ICU is enabled by the user, this still takes precedence.
Also, note that MinGW-w64 misses the respective Windows API, and will
therefore still use the old fallback.

[ChangeLog][QtCore][Text] Qt on Windows now uses the ICU API of the
Windows SDK for text codec support (if available and no 'full' ICU is
configured). This greatly enhances the list of supported text codecs
available.

Fixes: QTBUG-132056
Change-Id: Id02ae9f8cf56dee8c34795c0e60d2113f319bfd9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>