QTemporaryFile: add renameOverwrite()
... as public API for QTemporaryFileEngine::renameOverwrite().

This operation (to atomically rename a temp file to the final file) is
a building block in various safe I/O patterns and QTemporaryFile
should therefore have public API for it, not just a private one.

Adding similar functionality to QFile is orthogonal to this change, and
a lot more elaborate, since its rename() already overwrites, always,
even if non-atomic.

QTemporaryFile, OTOH, already guarantees atomicity. But its rename()
never overwrites (except when it does (QTBUG-132646)). So it's the
natural place to add this functionality, and helps write platform-
independent code in the face of QTBUG-132646.

[ChangeLog][QtCore][QTemporaryFile] Added renameOverwrite().

Fixes: QTBUG-132617
Fixes: QTBUG-132646
Task-number: QTBUG-132646
Task-number: QTBUG-2082
Change-Id: I405f554335005ec8c6f1762385a19ed3fac51217
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>