QFactoryLoader: move QPluginParsedMetaData to qplugin_p.h
Which is a more central header, with no other dependencies. This is just
a declarative structure with no .cpp requirements, so it fits there.

This way, qlibrary_p.h won't depend on qfactory_p.h. QLibrary,
QPluginLoader and QFactoryLoader are inextricably linked (the
plugin-parsing code is in qfactoryloader.cpp because QFactoryLoader
needs it to parse metadata too). This will later allow me to make
qfactory_p.h depend on qlibrary_p.h.

Change-Id: I0544826807f97d366142fffde0ebdc802273ffc2
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>