QFactoryLoader: unload the plugins we loaded
We just left everything for QLibraryStore::cleanup() to do when that ran
at QtCore unload time. Which is certainly weird because that meant no
other library could be unloaded if it was a dependency of a plugin we'd
loaded. Now, this should allow the libraries to start unloading before
QtCore, when the Q_GLOBAL_STATICs holding the QFactoryLoaders get
destroyed by ::exit().

This also explicitly deletes the instances in each of those dynamic
plugins, otherwise they may remain around and cause crashes later. This
is not done for instances coming from staticplugins - see next commit.

Change-Id: I752d41069e192c7be4a0fffd5ab0a253108b0b0c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>